3_AIVision/ocr_use_aivision/requirements.txt
azure-core
azure-ai-vision-imageanalysis